commands: doom: import doomgeneric source port
doomgeneric[1] is a fork of fbDOOM wit the intention on making DOOM easier
to port. It still assumes a standard C library to be available, but
it saves us the hassle of refactoring the X11 drawing routines.

The port lacks sound, but this will be retrofitted in a later commit.

[1]: https://github.com/ozkl/doomgeneric

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>