aiodev: add ST L3GD20 Gyroscope and temperature sensor support
The L3GD0 is an on-board peripheral of the STM32F429-DISCOVERY and as
such a good target to test out SPI driver support. Add a simple aiodev
driver to interface with it. It can read the chip's angular rate in
three dimensions as well as the temperature sensor embedded in it.

Signed-off-by: Ahmad Fatoum <ahmad@a3f.at>